Split TimeTest into 2 separate tests, 1 for testing options and other
... for testing the Time class
